requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
psycopg2-binary>=2.9
django-environ>=0.10
gunicorn==22.*
